projects/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.11.3
pytest==8.3.5
python-slugify==8.0.4
requests==2.32.3
SQLAlchemy==2.0.40
text-unidecode==1.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.3.0
Werkzeug==3.1.3
flask-marshmallow==1.3.0
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.2
flasgger==0.9.7.1
psycopg2-binary
